Split out master/non-master suites
Make the default suite a non-master setup and master the actual master setup.
Adjusted the tests to reflect that. There technically should be a noded service
running on the non-master setup but since its not a part of a cluster, it won't
be running yet. So for now its rather difficult to test that.

Also move the data bags and encrypted key into a common directory for any suite.